feat: move components and utils into shared locations
