+++ web/src/features/alerts/pages/alerts-page/AggregateAlertTableRow.tsx
@@ -5,7 +5,8 @@ import { useTranslation } from 'react-i18next';
 import { getRuleUrl, usePerspective } from '../../../../shared/hooks/usePerspective';
 import { useMonitoringNamespace } from '../../../../shared/hooks/useMonitoringNamespace';
 import { AggregatedAlert } from './AlertsAggregates';
-import { AlertState, SeverityBadge } from '../../components/AlertUtils';
+import { AlertState } from '../../components/AlertUtils';
+import { SeverityBadge } from '../../../../shared/components/SeverityBadge';
 import AlertTableRow from './AlertTableRow';
 import { RuleResource } from '../../../../shared/utils/utils';
 import { Link } from 'react-router';
