+++ web/src/features/legacy-dashboards/components/dashboard-skeleton-legacy.tsx
@@ -4,7 +4,7 @@ import { memo, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Divider, PageSection, Split, SplitItem, Stack, StackItem } from '@patternfly/react-core';
-import { CombinedDashboardMetadata } from '../../perses-dashboards/hooks/useDashboardsData';
+import type { CombinedDashboardMetadata } from '../../../shared/types/types';
 import { DashboardDropdown } from '../../../shared/components/dashboard-dropdown';
 import { PollIntervalDropdown, TimespanDropdown } from './time-dropdowns';
 import { LegacyDashboardsAllVariableDropdowns } from './legacy-variable-dropdowns';
