+++ web/src/features/perses-dashboards/pages/dashboard-list-page/dashboard-list-frame.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactNode } from 'react';
 import { DashboardListHeader } from '../../components/dashboard-header';
-import { CombinedDashboardMetadata } from '../../hooks/useDashboardsData';
+import type { CombinedDashboardMetadata } from '../../../../shared/types/types';
 import { ProjectBar } from '../../components/project/ProjectBar';
 
 interface DashboardListFrameProps {
