+++ web/src/features/perses-dashboards/pages/dashboard-page/dashboard-frame.tsx
@@ -1,7 +1,7 @@
 import { FC, ReactNode } from 'react';
 import { DashboardEmptyState } from '../../components/emptystates/DashboardEmptyState';
 import { DashboardHeader } from '../../components/dashboard-header';
-import { CombinedDashboardMetadata } from '../../hooks/useDashboardsData';
+import type { CombinedDashboardMetadata } from '../../../../shared/types/types';
 import { ProjectBar } from '../../components/project/ProjectBar';
 import { PersesWrapper } from '../../components/PersesWrapper';
 import { ToastProvider } from '../../components/ToastProvider';
