+++ web/src/features/perses-dashboards/pages/dashboard-page/dashboard-toolbar.tsx
@@ -16,13 +16,13 @@ import {
 import { TimeRangeControls, useTimeZoneParams } from '@perses-dev/plugin-system';
 import { ReactElement, ReactNode, useCallback, useEffect } from 'react';
 
-import { useAccessReview } from '@openshift-console/dynamic-plugin-sdk';
 import { StackItem } from '@patternfly/react-core';
 import * as _ from 'lodash-es';
 import PencilIcon from 'mdi-material-ui/PencilOutline';
 import { useTranslation } from 'react-i18next';
 import { DashboardDropdown } from '../../../../shared/components/dashboard-dropdown';
 import { useDashboardsData } from '../../hooks/useDashboardsData';
+import { usePersesEditPermissions } from '../../hooks/usePersesEditPermissions';
 
 import { persesDashboardDataTestIDs } from '../../../../shared/constants/data-test';
 
@@ -98,34 +98,6 @@ export const EditButton = ({ onClick, activeProject }: EditButtonProps): ReactEl
   }
 
   return button;
-};
-
-export const usePersesEditPermissions = (namespace: string | null = null) => {
-  const [canCreate, createLoading] = useAccessReview({
-    group: 'perses.dev',
-    resource: 'persesdashboards',
-    verb: 'create',
-    namespace,
-  });
-
-  const [canUpdate, updateLoading] = useAccessReview({
-    group: 'perses.dev',
-    resource: 'persesdashboards',
-    verb: 'update',
-    namespace,
-  });
-
-  const [canDelete, deleteLoading] = useAccessReview({
-    group: 'perses.dev',
-    resource: 'persesdashboards',
-    verb: 'delete',
-    namespace,
-  });
-
-  const loading = createLoading || updateLoading || deleteLoading;
-  const canEdit = canUpdate && canCreate && canDelete;
-
-  return { canEdit, loading };
 };
 
 export const OCPDashboardToolbar = (props: DashboardToolbarProps): ReactElement => {
