+++ web/src/shared/types/types.ts
@@ -1,4 +1,5 @@
 import { Alert, PrometheusLabels, Silence } from '@openshift-console/dynamic-plugin-sdk';
+import type { DashboardResource } from '@perses-dev/core';
 
 export const enum AlertSource {
   Platform = 'platform',
@@ -58,4 +59,12 @@ export type PatternflyToken = {
   name: string;
   value: string;
   var: string;
+};
+
+export type CombinedDashboardMetadata = {
+  name: string;
+  project?: string;
+  tags: string[];
+  title: string;
+  persesDashboard?: DashboardResource;
 };
